Updated api.py, sync.py, and update.py update.py Removed unused imports.
sync.py
Added debugging statements. And made a new function to check if the entries are valid before the hidden counter
is updated from 0 to 1, then 2. For more information about the hidden counter refer to README.md.

The functions are schedule and rowCount.

api.py
added a new argument in the url. The purpose is to keep track of which department has accessed the api.
Currently all it does is print out which department is currently accessing the api.