Changes made to api.py details below TLDR; SQL query made more streamlined, and is now less error prone. Made checks dynamic. Further reducing its tendancy to error out.
Detailed:
Previously when using the unmodified SQL query. There would be an extra value within it. And further down would be a
list comprehension with unpacking. To amend it slices would be used to delete that extra unwanted value. And there wasn't
a method to determine which position it was dynamically. Without large memory costs. Now a more memory safe method is used.

The current method used now is that the datetime module is used, along with a nested for loop to determine which element,
is a datetime.date object. Then format it with a method availble to all datetime.date objects called strftime(). Then, pass
in the format that is appropriate. Therefore, instead of having it be statically determined. As in the future, the size of
the table in terms of columns might increase. And this will reduce the tendancy of errors if it does occur.