Detailed changes made below. Global changes to all api resources: A recently discovered issue has been fixed: When a check fails it will have a return statement, to end the fucntion. That was the previous assumption. However, this is not the case. And has been fixed with the abort function from flask_restful. This change is made globally to all api resources files. Previously, the host, password, and other valid information needed to connect to a database needs to be changed manually when testing needs to be done on the sandbox MySQL table. So it has been made more convinient via lists, and indexes.
1. api.py
The need for a key to access to API has been temporarily removed, until a method of generating a key is determined.
A few changes has been made, regarding method of connection to a database.

2. sync.py
Visual change: a very long insert statement has been broken down further, to reduce the amount of space it takes on the screen.
The verify function has been removed. While testing, the inclusion or exclusion of the verify function makes no difference. Which
is why the function has been removed.

Order of commit statements have been rearranged due to issues where the needed changes are overwritten.

A new module:
A new module has been added to QA. As a way of verifying wheter an api key is legitmate or not. Development on this is slow, as
a method of giving a new user his/her unique api key is still undefined.