Changes listed below
Global changes to all endpoints:
1. The way to connecto to a SQL database which previously use the reading of files have been changed. It is now hard-coded. For anyone reading the
code in the future. The database, will be deleted, and the user will be deleted as well. Therefore, trying to get a look at the password is
useless.

2. All end points have been moved to env-api/endpoints. Therefore the folder has been made available, and no longer in the .gitignore.

Removed:
1. lookup.py it is seen as unnecessary as spec.py has been added, since it is a more advanced version of lookup.py
2. update.py has been removed as well. Since it is no longer in use.

Minor changes:
codecs has been used in api.py to render a html file. That contains a simple greet page.